# gh-91404: Improve `re` performance