--- a/Makefile.pre.in
+++ b/Makefile.pre.in
@@ -1356,7 +1356,7 @@ regen-sre:
 	$(PYTHON_FOR_REGEN) $(srcdir)/Tools/scripts/generate_sre_constants.py \
 		$(srcdir)/Lib/re/_constants.py \
 		$(srcdir)/Modules/_sre/sre_constants.h \
-		$(srcdir)/Modules/_sre/sre_targets.h \
+		$(srcdir)/Modules/_sre/sre_targets.h
 
 Python/compile.o Python/symtable.o Python/ast_unparse.o Python/ast.o Python/future.o: $(srcdir)/Include/internal/pycore_ast.h