… Lib/re/_constants.py
… Modules/_sre/sre_targets.h