@markshannon 2022-04-13
Nice speedup.

I wonder how much of the speedup would be achieved by just moving `ctx->pattern` and `ctx->ptr` into local variables and leaving the dispatch alone. I'm wondering this for a couple of reasons:

1. [This paper](https://hal.inria.fr/hal-01100647/document) suggests that it is the memory read, not the branch prediction, that has the larger impact on performance
2. We are concerned about [dispatching on Windows](https://github.com/python/cpython/issues/89279), and it might be good to apply the results of that here as well.

How easy would it be to just apply the changes that move `ctx->pattern` and `ctx->ptr`, without the dispatching changes?